requirements.txt
pillow==10.4.0
feedparser==6.0.11
requests==2.32.3